+++ packages/gatsby-recipes/src/providers/utils/get-graphql-fields.js
@@ -1,7 +1,5 @@
-const {
-  makeExecutableSchema,
-  SchemaDirectiveVisitor,
-} = require(`graphql-tools`)
+const { makeExecutableSchema } = require(`@graphql-tools/schema`)
+const { SchemaDirectiveVisitor } = require(`@graphql-tools/utils`)
 
 const gqlFieldsToArray = fields =>
   Object.entries(fields).reduce((acc, [key, value]) => {