+++ packages/gatsby/cache-dir/static-entry.js
@@ -17,7 +17,7 @@ const {
 
 const { RouteAnnouncerProps } = require(`./route-announcer-props`)
 const apiRunner = require(`./api-runner-ssr`)
-const syncRequires = require(`./sync-requires`)
+const syncRequires = require(`$virtual/sync-requires`)
 const { version: gatsbyVersion } = require(`gatsby/package.json`)
 
 const stats = JSON.parse(
@@ -60,6 +60,8 @@ const getPageDataUrl = pagePath => {
   const pageDataPath = getPageDataPath(pagePath)
   return `${__PATH_PREFIX__}/${pageDataPath}`
 }
+
+const getStaticQueryUrl = hash => `${__PATH_PREFIX__}/static/d/${hash}.json`
 
 const getPageData = pagePath => {
   const pageDataPath = getPageDataPath(pagePath)
@@ -201,7 +203,9 @@ export default (pagePath, callback) => {
 
   const appDataUrl = getAppDataUrl()
 
-  const { componentChunkName } = pageData
+  const { componentChunkName, staticQueryHashes = [] } = pageData
+
+  const staticQueryUrls = staticQueryHashes.map(getStaticQueryUrl)
 
   class RouteHandler extends React.Component {
     render() {
@@ -361,6 +365,18 @@ export default (pagePath, callback) => {
       />
     )
   }
+  staticQueryUrls.forEach(staticQueryUrl =>
+    headComponents.push(
+      <link
+        as="fetch"
+        rel="preload"
+        key={staticQueryUrl}
+        href={staticQueryUrl}
+        crossOrigin="anonymous"
+      />
+    )
+  )
+
   if (appDataUrl) {
     headComponents.push(
       <link
@@ -432,17 +448,29 @@ export default (pagePath, callback) => {
     />
   )
 
+  let bodyScripts = []
+  if (chunkMapping[`polyfill`]) {
+    chunkMapping[`polyfill`].forEach(script => {
+      const scriptPath = `${__PATH_PREFIX__}${script}`
+      bodyScripts.push(
+        <script key={scriptPath} src={scriptPath} noModule={true} />
+      )
+    })
+  }
+
   // Filter out prefetched bundles as adding them as a script tag
   // would force high priority fetching.
-  const bodyScripts = scripts
-    .filter(s => s.rel !== `prefetch`)
-    .map(s => {
-      const scriptPath = `${__PATH_PREFIX__}/${JSON.stringify(s.name).slice(
-        1,
-        -1
-      )}`
-      return <script key={scriptPath} src={scriptPath} async />
-    })
+  bodyScripts = bodyScripts.concat(
+    scripts
+      .filter(s => s.rel !== `prefetch`)
+      .map(s => {
+        const scriptPath = `${__PATH_PREFIX__}/${JSON.stringify(s.name).slice(
+          1,
+          -1
+        )}`
+        return <script key={scriptPath} src={scriptPath} async />
+      })
+  )
 
   postBodyComponents.push(...bodyScripts)