+++ packages/gatsby/src/commands/types.ts
@@ -1,8 +1,15 @@
 import { PackageJson, Reporter } from "gatsby"
+import { Store, AnyAction } from "redux"
+import { IGatsbyState } from "../redux/types"
 
 export interface ICert {
-  key: string
-  cert: string
+  key: Buffer
+  cert: Buffer
+}
+
+export interface IDebugInfo {
+  port: number
+  break: boolean
 }
 
 export interface IProgram {
@@ -21,6 +28,10 @@ export interface IProgram {
   https?: boolean
   sitePackageJson: PackageJson
   ssl?: ICert
+  inspect?: number
+  inspectBrk?: number
+  graphqlTracing?: boolean
+  setStore?: (store: Store<IGatsbyState, AnyAction>) => void
 }
 
 // @deprecated