+++ www/src/components/guidelines/cards/starter.js
@@ -1,11 +1,18 @@
-import React from "react"
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 import { MdStar } from "react-icons/md"
 
-import { Box, Flex, Heading, Text } from "../system"
+import { Box, Flex } from "theme-ui"
+import { Heading, Text } from "../system"
 import ImagePlaceholder from "../image-placeholder"
 
 const Card = () => (
-  <Box width="20rem" mb={8}>
+  <Box
+    sx={{
+      width: `20rem`,
+      mb: 8,
+    }}
+  >
     <ImagePlaceholder borderRadius={1} boxShadow="raised" mb={2} />
     <Flex>
       <Text
@@ -13,13 +20,18 @@ const Card = () => (
         fontFamily="heading"
         fontSize={1}
         letterSpacing="tracked"
-        css={{
-          textTransform: `uppercase`,
-        }}
+        textTransform="uppercase"
       >
         greglobinski/
       </Text>
-      <Flex alignItems="center" color="grey.60" fontSize={1} ml="auto">
+      <Flex
+        sx={{
+          alignItems: `center`,
+          color: `grey.60`,
+          fontSize: 1,
+          ml: `auto`,
+        }}
+      >
         <MdStar /> 167
       </Flex>
     </Flex>
@@ -27,7 +39,11 @@ const Card = () => (
       gatsby-starter-personal-blog
     </Heading>
     <Text fontSize={1}>A Gatsby starter for a personal blog</Text>
-    <Flex mt={4}>
+    <Flex
+      sx={{
+        mt: 4,
+      }}
+    >
       <Text color="grey.50" fontSize={0}>
         Updated 3 days ago
       </Text>