+++ www/src/components/sidebar/sidebar.js
@@ -17,7 +17,7 @@ function isItemInActiveTree(item, activeItem, activeItemParents) {
 
 function getOpenItemHash(itemList, activeItem, activeItemParents) {
   let result = {}
-  for (let item of itemList) {
+  for (const item of itemList) {
     if (item.items) {
       result[item.title] = isItemInActiveTree(
         item,
@@ -100,7 +100,7 @@ export default function Sidebar({
   const getItemState = React.useCallback(
     item => {
       return {
-        isExpanded: openSectionHash[item.title] || disableAccordions,
+        isExpanded: openSectionHash[item.title] || disableAccordions || false,
         isActive: item.title === activeItem.title,
         inActiveTree: isItemInActiveTree(item, activeItem, activeItemParents),
       }
@@ -150,6 +150,7 @@ export default function Sidebar({
             border: 0,
             display: `block`,
             overflowY: `auto`,
+            overscrollBehavior: `contain`,
             transition: t =>
               `opacity ${t.transition.speed.default} ${t.transition.curve.default}`,
             zIndex: 10,