+++ packages/gatsby-plugin-react-helmet/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.9...gatsby-plugin-react-helmet@3.3.10) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.8...gatsby-plugin-react-helmet@3.3.9) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.7...gatsby-plugin-react-helmet@3.3.8) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.6...gatsby-plugin-react-helmet@3.3.7) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.5...gatsby-plugin-react-helmet@3.3.6) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.4...gatsby-plugin-react-helmet@3.3.5) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.3...gatsby-plugin-react-helmet@3.3.4) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.2...gatsby-plugin-react-helmet@3.3.3) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
 ## [3.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-react-helmet@3.3.1...gatsby-plugin-react-helmet@3.3.2) (2020-05-20)
 
 **Note:** Version bump only for package gatsby-plugin-react-helmet